refactor reporting job submission routes
